apps/connector/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
python-json-logger==3.2.1
PyYAML==6.0.2
jsonschema==4.23.0
